+++ examples/test-smtp-connection-with-callbacks.js
@@ -1,23 +1,26 @@
 'use strict'
 
+const fs = require('fs')
+const MailComposer = require('nodemailer/lib/mail-composer')
+
 const SMTPConnection = require('nodemailer/lib/smtp-connection')
 
-const fs = require('fs')
-const mailcomposer = require('mailcomposer')
-
 // Usage: node test-smtp-client.js host=localhost port=25 ignoreTLS=true user=u pass=p from=a@example.com to=b@example.net data=-
-const options = Object.assign({
+const defaultOptions = {
   opportunisticTLS: true,
   from: 'sender@example.com',
   to: 'recpient@example.net'
-}, ...process.argv.slice(2).map(a => a.split('=')).map(([k, v]) => ({[k]: v})))
-const {from, to, user, pass} = options
+}
+const userOptions = Object.assign({}, ...process.argv.slice(2).map((a) => a.split('=')).map(([k, v]) => ({ [k]: v })))
+
+const options = { ...defaultOptions, ...userOptions }
+const { from, to, user, pass } = options
 
 const message = options.data === '-' ? process.stdin
-  : !options.data ? mailcomposer({from, to}).createReadStream()
+  : !options.data ? new MailComposer({ from, to }).compile().createReadStream()
   : fs.readFileSync(options.data)
 
-const envelope = {from, to}
+const envelope = { from, to: [to] }
 
 const connection = new SMTPConnection(options)